Move NewRestfulDasServerOnRandomPort into _test file
